[Ref] Refactor frontend components.
