+++ frontend/src/components/match/LabelEditDialog.vue
