+++ frontend/src/components/match/TransformationSummaryView.vue
